+++ src/main/java/dk/emilvn/exam/result/ResultDTO.java
@@ -2,5 +2,5 @@ package dk.emilvn.exam.result;
 
 import java.time.LocalDate;
 
-public record ResultDTO(Long id, LocalDate date, double result, ResultType resultType, Long participantId, Long disciplineId) {
+public record ResultDTO(Long id, LocalDate date, int result, ResultType resultType, Long participantId, Long disciplineId) {
 }
